requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
tqdm>=4.66
numpy>=1.24
pillow>=10.0
pytest>=7.0
PySide6>=6.7.0
